+++ R/get_add.R
@@ -52,8 +52,9 @@
 #' only tables with actual *df* change
 #' equal to expected *df* change.
 #'
-#' @return A named list of parameter
-#' tables to be used by
+#' @return An object of the class
+#' `partables`, a named list of parameter
+#' tables, each of them to be used by
 #' [lavaan::lavaan()] or [update()]
 #' for fitting a model with the added
 #' parameters.
@@ -151,6 +152,7 @@ get_add <- function(sem_out,
         paste("add:", out)
       })
     names(out) <- out_names
+    class(out) <- c("partables", class(out))
     out
   }